+++ tests/api/test_search.py
@@ -150,6 +150,12 @@ def test_search_geometry_bbox():
         (100.0, 1.0, 105.0, 0.0),  # ymin greater than ymax
         (100.0, 0.0, 5.0, 105.0, 1.0, 4.0),  # min elev greater than max elev
         (-200.0, 0.0, 105.0, 1.0),  # xmin is invalid WGS84
+        (
+            105.0,
+            0.0,
+            100.0,
+            1.0,
+        ),  # xmin greater than xmax but not crossing Antimeridian
         (100.0, -100, 105.0, 1.0),  # ymin is invalid WGS84
         (100.0, 0.0, 190.0, 1.0),  # xmax is invalid WGS84
         (100.0, 0.0, 190.0, 100.0),  # ymax is invalid WGS84